create configmap with values
