+++ deploy/operator.yaml
@@ -31,6 +31,11 @@ spec:
                   fieldPath: metadata.name
             - name: OPERATOR_NAME
               value: "ibm-iam-operator"
+            - name: ICP_CONSOLE_URL
+              valueFrom:
+                configMapKeyRef:
+                  name: management-ingress-info
+                  key: MANAGEMENT_INGRESS_ROUTE_HOST
           volumeMounts:
             - mountPath: "/certs"
               name: cluster-ca-cert
