+++ pkg/controller/authentication/configmap.go
@@ -25,23 +25,30 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"os"
 	"strconv"
+	"strings"
 )
 
-func (r *ReconcileAuthentication) handleConfigMap(instance *operatorv1alpha1.Authentication, currentConfigMap *corev1.ConfigMap, requeueResult *bool) error {
+func (r *ReconcileAuthentication) handleConfigMap(instance *operatorv1alpha1.Authentication, wlpClientID string, wlpClientSecret string, currentConfigMap *corev1.ConfigMap, requeueResult *bool) error {
 
-	configMapList := []string{"platform-auth-idp", "registration-json", "registration-script"}
+	configMapList := []string{"platform-auth-idp", "registration-script", "registration-json", }
 
-	functionList := []func(*operatorv1alpha1.Authentication, *runtime.Scheme) *corev1.ConfigMap{authIdpConfigMap, registrationJsonConfigMap, registrationScriptConfigMap}
+	functionList := []func(*operatorv1alpha1.Authentication, *runtime.Scheme) *corev1.ConfigMap{authIdpConfigMap, registrationScriptConfigMap}
 
 	reqLogger := log.WithValues("Instance.Namespace", instance.Namespace, "Instance.Name", instance.Name)
 	var err error
+	var newConfigMap *corev1.ConfigMap
 
 	for index, configMap := range configMapList {
 		err = r.client.Get(context.TODO(), types.NamespacedName{Name: configMap, Namespace: instance.Namespace}, currentConfigMap)
 		if err != nil && errors.IsNotFound(err) {
 			// Define a new ConfigMap
-			newConfigMap := functionList[index](instance, r.scheme)
+			if configMapList[index] == "registration-json" {
+				newConfigMap = registrationJsonConfigMap(instance, wlpClientID, wlpClientSecret, r.scheme)
+			} else {
+				newConfigMap = functionList[index](instance, r.scheme)
+			}
 			reqLogger.Info("Creating a new ConfigMap", "ConfigMap.Namespace", instance.Namespace, "ConfigMap.Name", configMap)
 			err = r.client.Create(context.TODO(), newConfigMap)
 			if err != nil {
@@ -139,8 +146,14 @@ func authIdpConfigMap(instance *operatorv1alpha1.Authentication, scheme *runtime
 	return newConfigMap
 }
 
-func registrationJsonConfigMap(instance *operatorv1alpha1.Authentication, scheme *runtime.Scheme) *corev1.ConfigMap {
+func registrationJsonConfigMap(instance *operatorv1alpha1.Authentication, wlpClientID string, wlpClientSecret string, scheme *runtime.Scheme) *corev1.ConfigMap {
 	reqLogger := log.WithValues("Instance.Namespace", instance.Namespace, "Instance.Name", instance.Name)
+    icpConsoleURL := os.Getenv("ICP_CONSOLE_URL")
+	strings.ReplaceAll(registrationJson, "WLP_CLIENT_ID", wlpClientID)
+	strings.ReplaceAll(registrationJson, "WLP_CLIENT_SECRET", wlpClientSecret)
+	strings.ReplaceAll(registrationJson, "ICP_CONSOLE_URL", icpConsoleURL)
+
+
 	newConfigMap := &corev1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "registration-json",
