+++ pkg/controller/authentication/constants.go
@@ -29,9 +29,6 @@ if [ $HTTP_CODE = "404" ]
 then
   echo "Running new client id registration"
   cp /jsons/platform-oidc-registration.json platform-oidc-registration.json
-  sed -i "s/WLP_CLIENT_ID/$WLP_CLIENT_ID/g" platform-oidc-registration.json
-  sed -i "s/WLP_CLIENT_SECRET/$WLP_CLIENT_SECRET/g" platform-oidc-registration.json
-  sed -i "s/ICP_CONSOLE_URL/$ICP_CONSOLE_URL/g" platform-oidc-registration.json
   until curl -i -k -X POST -u oauthadmin:$WLP_CLIENT_REGISTRATION_SECRET \
    -H "Content-Type: application/json" \
    --data @platform-oidc-registration.json \
@@ -39,9 +36,6 @@ then
 else
   echo "Running update client id registration."
   cp /jsons/platform-oidc-registration.json platform-oidc-registration.json
-  sed -i "s/WLP_CLIENT_ID/$WLP_CLIENT_ID/g" platform-oidc-registration.json
-  sed -i "s/WLP_CLIENT_SECRET/$WLP_CLIENT_SECRET/g" platform-oidc-registration.json
-  sed -i "s/ICP_CONSOLE_URL/$ICP_CONSOLE_URL/g" platform-oidc-registration.json
   until curl -i -k -X PUT -u oauthadmin:$WLP_CLIENT_REGISTRATION_SECRET \
    -H "Content-Type: application/json" \
    --data @platform-oidc-registration.json \
@@ -49,7 +43,7 @@ else
 fi
 `
 
-const registrationJson = `{
+var registrationJson string = `{
    "token_endpoint_auth_method":"client_secret_basic",
    "client_id": "WLP_CLIENT_ID",
    "client_secret": "WLP_CLIENT_SECRET",
